+++ changelog.md
@@ -1,2 +1,6 @@
+# v1.0.1
+- Added clear search button
+- Set default value for setting *`Clear Search on Refresh`* to `false`
+
 # v1.0.0
 - Initial release